Import of LPRng 3.6.11
git-svn-id: svn+ssh://svn.mit.edu/athena/branches/vendor@13798 728af825-273c-0410-89f9-f7d3b574a069

+++ third/lprng/lprng/src/include/lpd_rcvjob.h
@@ -4,7 +4,7 @@
  * Copyright 1988-1999, Patrick Powell, San Diego, CA
  *     papowell@astart.com
  * See LICENSE for conditions of use.
- * $Id: lpd_rcvjob.h,v 1.1.1.1 1999-05-04 18:07:14 danw Exp $
+ * $Id: lpd_rcvjob.h,v 1.1.1.2 1999-10-27 20:10:14 mwhitson Exp $
  ***************************************************************************/
 
 
@@ -17,7 +17,7 @@ int Receive_job( int *sock, char *input );
 int Receive_block_job( int *sock, char *input );
 int Scan_block_file( int fd, char *error, int errlen );
 int Read_one_line( int fd, char *buffer, int maxlen );
-int Check_space( int jobsize, int min_space, char *pathname );
+int Check_space( double jobsize, int min_space, char *pathname );
 int Do_perm_check( struct job *job, char *error, int errlen );
 int Check_for_missing_files( struct job *job, struct line_list *files,
 	char *error, int errlen );
